+++ inst/+arduinoio/lib/OctaveCoreLibrary.cpp
@@ -205,7 +205,7 @@ OctaveCoreLibrary::OctaveCoreLibrary (OctaveArduinoClass &oc)
     }
 }
 void
-OctaveCoreLibrary::commandHandler (uint8_t cmdID, uint8_t* data, uint8_t datasz)
+OctaveCoreLibrary::commandHandler (uint8_t cmdID, uint8_t* data, uint16_t datasz)
 {
   int val;
