+++ inst/+arduinoio/lib/OctaveSerialLibrary.cpp
@@ -80,7 +80,7 @@ OctaveSerialLibrary::setup ()
 }
 
 void
-OctaveSerialLibrary::commandHandler (uint8_t cmdID, uint8_t* data, uint8_t datasz)
+OctaveSerialLibrary::commandHandler (uint8_t cmdID, uint8_t* data, uint16_t datasz)
 {
   switch (cmdID)
     {
