+++ inst/+arduinoioaddons/+EEPRomAddon/EEPRomAddon.h
@@ -32,7 +32,7 @@ public:
     libName = "EEPRomAddon/EEProm";
     a.registerLibrary(this);
   }
-  void commandHandler(uint8_t cmdId, uint8_t* data, uint8_t datasz)
+  void commandHandler(uint8_t cmdId, uint8_t* data, uint16_t datasz)
   {
     switch(cmdId)
     {
