+++ inst/+arduinoioaddons/+ExampleLCD/LCDAddon.h
@@ -42,7 +42,7 @@ public:
   {
     lcd = 0;
   }
-  void commandHandler(uint8_t cmdId, uint8_t* data, uint8_t datasz)
+  void commandHandler(uint8_t cmdId, uint8_t* data, uint16_t datasz)
   {
     switch(cmdId)
     {
