+++ inst/+arduinoioaddons/+PulseTrain/src/PulseTrainAddon.h
@@ -23,6 +23,6 @@ class PulseTrainAddon : public LibraryBase
 public:
   PulseTrainAddon(OctaveArduinoClass& a);
 
-  void commandHandler(uint8_t cmdID, uint8_t* data, uint8_t datasz);
+  void commandHandler(uint8_t cmdID, uint8_t* data, uint16_t datasz);
   virtual void loop();
 };
