+++ inst/+arduinoioaddons/+RTCAddon/DS1307Addon.h
@@ -27,7 +27,7 @@ public:
     libName = "RTCAddon/DS1307";
     a.registerLibrary(this);
   }
-  void commandHandler(uint8_t cmdId, uint8_t* data, uint8_t datasz)
+  void commandHandler(uint8_t cmdId, uint8_t* data, uint16_t datasz)
   {
     switch(cmdId)
     {
