+++ inst/+arduinoioaddons/+adafruit/motorshieldv2.h
@@ -73,7 +73,7 @@ public:
     //
     //   servo1.attach(10);
   }
-  void commandHandler(uint8_t cmdId, uint8_t* data, uint8_t datasz)
+  void commandHandler(uint8_t cmdId, uint8_t* data, uint16_t datasz)
   {
     switch(cmdId)
     {
